LayoutTests/editing/pasteboard/drag-files-to-editable-element-as-attachment-expected.txt

If we drag files onto an editable area, then attachments should be inserted into the editable area.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS attachment.nodeName is "ATTACHMENT"
PASS attachment.nodeName is "ATTACHMENT"
PASS attachment.nodeName is "ATTACHMENT"
PASS fileNames is "foo bar baz "
PASS successfullyParsed is true

TEST COMPLETE
